Organize components into subfolders
