+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import DebugMenu from "@/components/DebugMenu";
+import DebugMenu from "@/components/debug/DebugMenu";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
