+++ src/components/menu/MainMenu.tsx
@@ -1,4 +1,4 @@
-
+// src/components/menu/MainMenu.tsx
 // src/components/map-layout/main-menu.tsx
 
 "use client";
